requirements/common.txt
Django~=3.1.1
djangorestframework
Pillow
django-cors-headers
